Changes to support referencing subnets by IP and netmask.
This change implements support for refrencing subnet based on the
IP and netmask in the form ${IP10_0_1_0_24_address_0}. This
change also renames the get_address method to address to make it
more consistent with other method calls. If the subnet is referenced
by IP in the above form, but the subnet is not referenced in the
subnets yaml, the subnet will be created. This allows reference by IP
without a subnets yaml file. This change also makes the subnets
yaml file optional.